feat: Printer is now standalone
